Move `struct Holochain` into `mod holochain` and add `container_api` rustdoc description
